feat(ensadmin): introduce Next.js app (#316)
This PR introduces a new ensadmin-next app. It's going to replace the current ensadmin app once confirmed to work well as a Vercel deployment.

This PR aims to create foundations for further improvements. While reviewing this PR, please focus more on the foundations, and less about the features of the app. Please keep in mind, the a lot of functionality is going to be updated when #319 is completed.